+++ pkg/scripts_test/consts.go
@@ -17,9 +17,10 @@ const (
 
 	systemdDirectoryPath string = "/run/systemd/system"
 
-	installToken    string = "token"
-	installTokenEnv string = "SUMOLOGIC_INSTALL_TOKEN"
-	apiBaseURL      string = "https://open-collectors.sumologic.com"
+	installToken              string = "token"
+	installTokenEnv           string = "SUMOLOGIC_INSTALLATION_TOKEN"
+	deprecatedInstallTokenEnv string = "SUMOLOGIC_INSTALL_TOKEN"
+	apiBaseURL                string = "https://open-collectors.sumologic.com"
 
 	systemUser string = "otelcol-sumo"
 )